Refactor: InfinitiScroll de TSX para VUE
